Moved some files into src/exec/cmd_path.c
