+++ Makefile
@@ -8,7 +8,7 @@ RM = rm -rf
 SRC_MSH	=	frees.c main.c prompt.c signals.c
 SRC_BLT	=	cd.c echo.c env.c exit.c export.c pwd.c unset.c
 SRC_ERR =	errors.c syntax_errors.c syntax_errors2.c
-SRC_EXE =	exec_built_ins.c redirections.c heredoc.c run.c
+SRC_EXE =	cmd_path.c exec_built_ins.c redirections.c heredoc.c run.c
 SRC_PRS	=	parse_counter.c parse_split.c parse_str.c parse.c
 SRC_STC	=	cmd_utils.c content.c init.c
 SRC_UTL	=	array_utl.c list_utl.c env_utl.c str_utl.c
